+++ src/actions/session.js
@@ -1,20 +1,7 @@
-import { API_URL } from 'constants/config'
-// import axios from 'axios'
-// import _ from 'lodash'
 import { SubmissionError } from 'redux-form'
-import { createSelector } from 'reselect'
 import { fetchUtils } from 'utils'
-
-// ------------------------------------
-// Constants
-// ------------------------------------
-export const LOG_USER_IN = 'LOG_USER_IN'
-export const LOG_USER_OUT = 'LOG_USER_OUT'
-export const CURRENT_USER_REQUEST = 'CURRENT_USER_REQUEST'
-
-// ------------------------------------
-// Actions
-// ------------------------------------
+import { API_URL } from 'constants/config'
+import { LOG_USER_IN, LOG_USER_OUT, CURRENT_USER_REQUEST } from 'constants/session'
 
 /*  This is a thunk, meaning it is a function that immediately
     returns a function for lazy evaluation. It is incredibly useful for
@@ -73,35 +60,3 @@ export const actions = {
   logUserIn,
   logUserOut
 }
-
-// ------------------------------------
-// Action Handlers
-// ------------------------------------
-const ACTION_HANDLERS = {
-  [LOG_USER_IN]: (state, action) => ({ ...state, user: action.payload, isLoading: false, initialized: true }),
-  [LOG_USER_OUT]: (state) => ({ ...state, user: null, isLoading: false, initialized: true }),
-  [CURRENT_USER_REQUEST]: (state) => ({ ...state, isLoading: true })
-}
-
-// ------------------------------------
-// Reducer
-// ------------------------------------
-const initialState = {
-  user: undefined,
-  isLoading: false,
-  initialized: false
-}
-
-export default function sessionReducer (state = initialState, action) {
-  const handler = ACTION_HANDLERS[action.type]
-
-  return handler ? handler(state, action) : state
-}
-
-// ------------------------------------
-// Selectors
-// ------------------------------------
-
-export const selectSession = state => state.session
-
-export const selectUser = createSelector(selectSession, (session) => session.user)
