+++ src/constants/session.js
@@ -1,0 +1,3 @@
+export const LOG_USER_IN = 'LOG_USER_IN'
+export const LOG_USER_OUT = 'LOG_USER_OUT'
+export const CURRENT_USER_REQUEST = 'CURRENT_USER_REQUEST'
