+++ src/selectors/config.js
@@ -1,0 +1,1 @@
+export const selectConfig = state => state.config
