+++ src/selectors/session.js
@@ -1,0 +1,5 @@
+import { createSelector } from 'reselect'
+
+export const selectSession = state => state.session
+
+export const selectUser = createSelector(selectSession, (session) => session.user)
