+++ src/store/reducers.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux'
 import configReducer from 'reducers/config'
 import locationReducer from './location'
-import sessionReducer from './session'
+import sessionReducer from 'reducers/session'
 import { reducer as formReducer } from 'redux-form'
 
 export const makeRootReducer = (asyncReducers) => {
